add bla-file to demonstrate how to integrate files
the bla-file packages is built via normal poky methods.
but its a debian pkg, and its accessible in elbeimage.

bla-file.bb uses INHIBIT_DEFAULT_DEPS = "1" because
it does not depend on a cross compiler, etc.

so add the simple bb file and include it in IMAGE_INSTALL
in simple-image.bbappend.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>